Can /write without being logged. WIP. Needs signup modal

+++ application/classes/Controller/Page.php
@@ -4,13 +4,12 @@ class Controller_Page extends Controller_Project {
 	
 	public $template = 'templates/write';
 	
-	public function before()
-	{
-		$this->require_login();
-		return parent::before();
-	}
-	
 	public function action_daynotfound() {}
+	
+	public function action_anonymouswriting()
+	{
+		
+	}
 	
 	public function action_write()
 	{
@@ -19,41 +18,48 @@ class Controller_Page extends Controller_Project {
 		
 		if($_POST && strlen(arr::get($_POST, 'morningpage',''))>0)
 		{
-			$content = arr::get($_POST, 'morningpage','');
-			if($page->type == 'page')
+			if(user::logged())
 			{
-				$page->content = $page->content().$content;
-			}
-			try
-			{
-			    $page->wordcount = str_word_count(strip_tags($page->content()));
-                if($page->type == 'autosave')
-                {
-                    $page->type = 'page';
-                }
-				if($page->wordcount > 750 && !(bool)$page->counted)
+				$content = arr::get($_POST, 'morningpage','');
+				if($page->type == 'page')
 				{
-					user::update_stats($content, $page);
-					$page->counted = 1;
+					$page->content = $page->content().$content;
 				}
-				$page->update();
-				/*$autosave = $page->get_autosave();
-				if($autosave)
+				try
 				{
-					$autosave->delete();
-				}				
-				if(!(bool)$page->counted)
-				{
-					
+				    $page->wordcount = str_word_count(strip_tags($page->content()));
+	                if($page->type == 'autosave')
+	                {
+	                    $page->type = 'page';
+	                }
+					if($page->wordcount > 750 && !(bool)$page->counted)
+					{
+						user::update_stats($content, $page);
+						$page->counted = 1;
+					}
+					$page->update();
+					/*$autosave = $page->get_autosave();
+					if($autosave)
+					{
+						$autosave->delete();
+					}				
+					if(!(bool)$page->counted)
+					{
+						
+					}
+	                $page->save();*/
+					notes::success('Your page has been saved!');
+					site::redirect('write/'.$page->day);
 				}
-                $page->save();*/
-				notes::success('Your page has been saved!');
-				site::redirect('write/'.$page->day);
+				catch(ORM_Validation_Exception $e)
+				{
+					$errors = $e->errors('models');
+				}
 			}
-			catch(ORM_Validation_Exception $e)
+			else
 			{
-				$errors = $e->errors('models');
-			}
+				notes::error('You must be logged in to save your page.');
+			}	
 		}
 		$this->bind('errors', $errors);
 		$this->bind('page', $page);
